perf annotate: allow 's' on source code lines
In perf annotate, when 's' is pressed on a line containing
source code, it shows the message "Only available for assembly
lines".
This patch gets rid of the error, moving the cursr to the next
available asm line (or the closest previous one if no asm line
is found moving forwards), before hiding source code lines.

Signed-off-by: Riccardo Mancini <rickyman7@gmail.com>