perf: fix segfault when wrong option for --debug is provided
The command `perf --debug record` causes a segfault instead of notifying
the user of the unrecognized option for --debug (the correct syntax would
be `perf --debug verbose=2 record`).

The bug is caused by a call to pr_err before debug_file (debug.c:34) is
set via perf_debug_setup(). It has been introduced in
commit 8abceacff87d ("perf debug: Add debug_set_file function").

This patch resolves the bug by calling perf_debug_setup earlier in the
main, so that debug_file is set to stderr before any call to a debug
printing function is made.

Signed-off-by: Riccardo Mancini <rickyman7@gmail.com>