perf: add read lock in maps__first
maps__first does not take a read lock, which could lead to concurrency issues
with maps__empty.
This patch adds the read lock to maps__first and adds a new variant
__maps__first which does not take the lock and is used in looping macros.
In addition, map__next is renamed to __map__next for coherency with __maps_first,
to emphasize that it requires the lock on the parent maps.