perf: have maps__find grab a refcount on map while holding the lock
Before this patch, there could have been race conditions between a thread
deleting a map entry from the list and one searching the list and returning
an element.
I.e. if thread A had a reference to a map from maps__find (without the safe
refcnt), then thread B could delete that map from the list and invalidate
the reference of thread A.
This patch adds a map__get in the safe section of maps__find so that this
condition cannot happen. In addition, this patch adds the corresponding
map__put when the function is done with the reference.
This is part of a patchset, therefore some functions will be dealt with later,
in these cases a map__put is called to maintain the status quo.