perf: propagate refcnt'ed map from maps__find_ams
This patch changes maps__find_ams to handle the refcnt from maps__find.
maps__find_ams has an input/output argument ams which already contains a
map. This change will make it mandatory to hold a refcount on the map, and
it will release this refcnt if the map is to be replaced with another (refcnted)
map. In addition, corresponding map__put are added to callers.