perf: add refcounts to addr_location members
Following the previous change of returning a refcounted map from maps__find, this
patch adds a refcount also to other members of the addr_location struct, i.e.
map (from maps__find), maps and thread. Callers will need to call addr_location__put_members
to release all of them.
Furthermore, since we're adding refcounting, we also need to make sure that also in the other
places where an addr_location struct is filled, these refcounts are taken.
The addr_location__put_members will be added in a subsequent patch for clarity
(this patch wil have memory leaks but it's not a big deal).