perf libc_compat: change COMPAT_NEED_REALLOCARRAY with HAVE_REALLOCARRAY
In libc_compat, the reallocarray feature check defines the
COMPAT_NEED_REALLOCARRAY macro, which means that the feature is not
present and, thus, the libc_compat.h header should define reallocarray.
However, this is in contrast to all other feature checks which define a
macro HAVE_* when the feature is present.
In order to keep uniformity with other feature checks and within
libc_compat.h, this patch changes the macro from COMPAT_NEED_REALLOCARRAY
to HAVE_REALLOCARRAY.

Furthermore, in perf's Makefile.config, said macro is defined only in a
specific case (CORESIGHT is defined and libopencsd is avilable).
Since this is a feature of glibc, this feature check should not be
specific to that case.
Therefore, this patch also moves it outside the current checks, alongside
all existing feature checks.
By doing so, reallocarray will be readily available to use in future perf
patches, in case it will ever be needed.

Signed-off-by: Riccardo Mancini <rickyman7@gmail.com>