perf symbols: protect dso->nsinfo with dso->lock
dso->nsinfo is a cache for thread->nsinfo but is accessed and modified
without being protected by a lock.
In perf-top, this leads to a race condition during the multi-threaded
synthesis of existing threads, in particular in the map__new function.
In this function the refcount of dso->nsinfo is decreased and its
reference is replaced by another one. Thus, what can happen is that the
refcount is decreased twice on the same reference. This makes the refcount
prematurely reach 0, causing a refcounting assertion error on exit, when
the dso is deleted. The frequency of this event is quite low.

This bug was not happening before the nsinfo refcounting fix patch
since the refcount was (wrongly) never decreased. Instead, this bug
surfaced the problem of the dso->nsinfo being concurrently accessed by
multiple threads.

This patch addresses this issue by protecting all accessess to the
dso->nsinfo with the dso->lock to prevent any possible race condition.

Testing on my laptop using perf stat to monitor perf-top shows no
noticeable performance loss, due to the low contention.

Fixes: 2d6b74baa7147251 ("perf map: Fix dso->nsinfo refcounting")
Signed-off-by: Riccardo Mancini <rickyman7@gmail.com>